move webhook config to internal/webhook
