+++ api/go.mod
@@ -19,7 +19,7 @@ require (
 
 require (
 	github.com/VictoriaMetrics/metrics v1.35.2 // indirect
-	github.com/VictoriaMetrics/metricsql v0.84.0 // indirect
+	github.com/VictoriaMetrics/metricsql v0.84.1 // indirect
 	github.com/aws/aws-sdk-go v1.55.6 // indirect
 	github.com/beorn7/perks v1.0.1 // indirect
 	github.com/bmatcuk/doublestar/v4 v4.8.1 // indirect
