+++ api/operator/v1beta1/vmalertmanager_types_test.go
@@ -1,19 +1,3 @@
-/*
-
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
 package v1beta1
 
 import (
@@ -40,7 +24,7 @@ global:
  resolve_timeout: 10m
  group_wait: 1s
           `
-			Expect(am.sanityCheck()).NotTo(Succeed())
+			Expect(am.Validate()).NotTo(Succeed())
 		})
 
 		It("Should allow with correct config syntax", func() {
@@ -57,7 +41,7 @@ global:
      # by default route to dev/null
      - name: blackhole
           `
-			Expect(am.sanityCheck()).To(Succeed())
+			Expect(am.Validate()).To(Succeed())
 		})
 	})
