+++ api/operator/v1beta1/vmnodescrape_types.go
@@ -57,12 +57,12 @@ type VMNodeScrapeList struct {
 }
 
 // AsProxyKey builds key for proxy cache maps
-func (cr VMNodeScrape) AsProxyKey() string {
+func (cr *VMNodeScrape) AsProxyKey() string {
 	return fmt.Sprintf("nodeScrapeProxy/%s/%s", cr.Namespace, cr.Name)
 }
 
 // AsMapKey - returns cr name with suffix for token/auth maps.
-func (cr VMNodeScrape) AsMapKey() string {
+func (cr *VMNodeScrape) AsMapKey() string {
 	return fmt.Sprintf("nodeScrape/%s/%s", cr.Namespace, cr.Name)
 }
