+++ api/operator/v1beta1/vmscrapeconfig_types.go
@@ -505,12 +505,12 @@ type VMScrapeConfigList struct {
 }
 
 // AsProxyKey builds key for proxy cache maps
-func (cr VMScrapeConfig) AsProxyKey(prefix string, i int) string {
+func (cr *VMScrapeConfig) AsProxyKey(prefix string, i int) string {
 	return fmt.Sprintf("scrapeConfigProxy/%s/%s/%s/%d", cr.Namespace, cr.Name, prefix, i)
 }
 
 // AsMapKey - returns cr name with suffix for token/auth maps.
-func (cr VMScrapeConfig) AsMapKey(prefix string, i int) string {
+func (cr *VMScrapeConfig) AsMapKey(prefix string, i int) string {
 	return fmt.Sprintf("scrapeConfig/%s/%s/%s/%d", cr.Namespace, cr.Name, prefix, i)
 }
