--- api/operator/v1beta1/vmsingle_webhook_test.go
@@ -1,47 +1,0 @@
-/*
-
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-package v1beta1
-
-import (
-	. "github.com/onsi/ginkgo/v2"
-)
-
-var _ = Describe("VMSingle Webhook", func() {
-
-	Context("When creating VMSingle under Validating Webhook", func() {
-		It("Should deny if a required field is empty", func() {
-
-			// TODO(user): Add your logic here
-
-		})
-
-		It("Should admit if all required fields are provided", func() {
-
-			// TODO(user): Add your logic here
-
-		})
-	})
-
-	Context("When creating VMSingle under Conversion Webhook", func() {
-		It("Should get the converted version of VMSingle", func() {
-
-			// TODO(user): Add your logic here
-
-		})
-	})
-
-})
