+++ config/rbac/kustomization.yaml
@@ -45,5 +45,3 @@ resources:
 # - operator_vmalert_viewer_role.yaml
 # - operator_vmagent_editor_role.yaml
 # - operator_vmagent_viewer_role.yaml
-#
-
