+++ internal/manager/manager.go
@@ -23,6 +23,7 @@ import (
 	"github.com/VictoriaMetrics/operator/internal/controller/operator/factory/k8stools"
 	"github.com/VictoriaMetrics/operator/internal/controller/operator/factory/logger"
 	"github.com/VictoriaMetrics/operator/internal/controller/operator/factory/reconcile"
+	webhookv1beta1 "github.com/VictoriaMetrics/operator/internal/webhook/operator/v1beta1"
 	"github.com/go-logr/logr"
 	promv1 "github.com/prometheus-operator/prometheus-operator/pkg/apis/monitoring/v1"
 	"github.com/prometheus-operator/prometheus-operator/pkg/apis/monitoring/v1alpha1"
@@ -47,7 +48,6 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/metrics"
 	metricsserver "sigs.k8s.io/controller-runtime/pkg/metrics/server"
 	"sigs.k8s.io/controller-runtime/pkg/webhook"
-	"sigs.k8s.io/controller-runtime/pkg/webhook/admission"
 	// +kubebuilder:scaffold:imports
 )
 
@@ -331,32 +331,26 @@ func RunManager(ctx context.Context) error {
 	return nil
 }
 
-type objectWithWebhookSetup interface {
-	runtime.Object
-	SetupWebhookWithManager(mgr ctrl.Manager) error
-	admission.CustomValidator
-}
-
 func addWebhooks(mgr ctrl.Manager) error {
-	f := func(objs []objectWithWebhookSetup) error {
-		for _, obj := range objs {
-			if err := obj.SetupWebhookWithManager(mgr); err != nil {
+	f := func(setupWebhooks []func(ctrl.Manager) error) error {
+		for _, setupWebhook := range setupWebhooks {
+			if err := setupWebhook(mgr); err != nil {
 				return err
 			}
 		}
 		return nil
 	}
-	return f([]objectWithWebhookSetup{
-		&vmv1beta1.VMAgent{},
-		&vmv1beta1.VMAlert{},
-		&vmv1beta1.VMSingle{},
-		&vmv1beta1.VMCluster{},
-		&vmv1beta1.VLogs{},
-		&vmv1beta1.VMAlertmanager{},
-		&vmv1beta1.VMAlertmanagerConfig{},
-		&vmv1beta1.VMAuth{},
-		&vmv1beta1.VMUser{},
-		&vmv1beta1.VMRule{},
+	return f([]func(ctrl.Manager) error{
+		webhookv1beta1.SetupVMAgentWebhookWithManager,
+		webhookv1beta1.SetupVMAlertWebhookWithManager,
+		webhookv1beta1.SetupVMSingleWebhookWithManager,
+		webhookv1beta1.SetupVMClusterWebhookWithManager,
+		webhookv1beta1.SetupVLogsWebhookWithManager,
+		webhookv1beta1.SetupVMAlertmanagerWebhookWithManager,
+		webhookv1beta1.SetupVMAlertmanagerConfigWebhookWithManager,
+		webhookv1beta1.SetupVMAuthWebhookWithManager,
+		webhookv1beta1.SetupVMUserWebhookWithManager,
+		webhookv1beta1.SetupVMRuleWebhookWithManager,
 	})
 }
